Worked on editors
Cleaned up the CreateModuleDialog which wasn't working before and
finished one of the TODOs there which was to update it to the new module
file system. I also fixed a lot of random bugs relating to that class
and how it behaved.

Two new editors were added as well: the ShellEditor and MessageEditor.
These now allow you to edit shell and message actions. More will be
written in the future.